Refactor: reducer 이름을 변경된 url에 맞게 수정
